Refactoring 2022 Tests (#33)
* Refactored Day 1 and 2 tests

* Refactored Day 3 tests

* Refactored Day 4 tests

* Refactored Day 5 Tests

* Refactored Day 8 tests

* Refactored Day 9 tests

* Refactored Day 10 tests

* Refactored Day 11 Tests

* Additional refactorings

+++ Problems/2022/Day11.cs
@@ -5,8 +5,8 @@ public class Day11
 
     public class Monkey
     {
-        private readonly Int128 worryFactor;
-        public Int128 Inspections = 0;
+        private readonly int worryFactor;
+        public Int64 Inspections = 0;
         private const string startingItemsString = "Starting items: ";
         private const string operationString = "Operation: new = old ";
         private const string divisibilityTestString = "Test: divisible by ";
@@ -19,18 +19,18 @@ public class Day11
             multiply,
             square
         }
-        public List<Int128> Items = new();
+        public List<Int64> Items = new();
         public Operation ItemOperation;
-        public Int128 OperationValue;
+        public int OperationValue;
         public int DivisibilityCheck;
         public (int trueMonkey, int falseMonkey) MonkeyToThrowTo;
 
-        public Monkey(string input, Int128 worryFactor)
+        public Monkey(string input, int worryFactor)
         {
             this.worryFactor = worryFactor;
             var inputLines = input.Split('\n', StringSplitOptions.TrimEntries);
 
-            Items = inputLines[0][startingItemsString.Length..].Split(',', StringSplitOptions.TrimEntries).Select(x => Int128.Parse(x)).ToList();
+            Items = inputLines[0][startingItemsString.Length..].Split(',', StringSplitOptions.TrimEntries).Select(x => Int64.Parse(x)).ToList();
 
             var operationLines = inputLines[1][operationString.Length..].Split(' ');
             switch (operationLines[1])
@@ -53,12 +53,12 @@ public class Day11
             MonkeyToThrowTo = (trueMonkey, falseMonkey);
         }
 
-        public List<(Int128 newItemValue, int destinationMonkey)> PerformInspections(int lcm) => Items.Select(x => Inspect(x, lcm)).ToList();
+        public List<(Int64 newItemValue, int destinationMonkey)> PerformInspections(int lcm) => Items.Select(x => Inspect(x, lcm)).ToList();
 
-        private (Int128 newItemValue, int destinationMonkey) Inspect(Int128 itemValue, int lcm)
+        private (Int64 newItemValue, int destinationMonkey) Inspect(Int64 itemValue, int lcm)
         {
             
-            Int128 newItemValue = itemValue;
+            Int64 newItemValue = itemValue;
             newItemValue = Operate(newItemValue);
             newItemValue /= worryFactor;
 
@@ -67,7 +67,7 @@ public class Day11
             return (newItemValue, (newItemValue % DivisibilityCheck == 0) ? MonkeyToThrowTo.trueMonkey : MonkeyToThrowTo.falseMonkey);
         }
 
-        private Int128 Operate(Int128 item) =>
+        private Int64 Operate(Int64 item) =>
         ItemOperation switch
         {
             Operation.square => item * item,
@@ -99,7 +99,7 @@ public class Day11
             Monkeys.Add(new Monkey(s[(s.IndexOf(':')+3)..], worryFactor));
     }
 
-    public Int128 Solve(int rounds)
+    public Int64 Solve(int rounds)
     {
         for (int round= 1; round <= rounds; round++)
         {
